+++ internal/model/model.go
@@ -1,4 +1,4 @@
-package common
+package model
 
 import "io"
